Separate current and old reservations and add availability toggle (#49)
* Remove hard coding when checking for times per day

* Add second tab for past reservations

* Separate past and future reservation in admin view

Split listing all reservations into two tabs, one for past and one for
future reservations. This is done to make it easier to hide past
reservations in the admin view.

* Add confirmation dialog when deleting reservations

Added a confirmation prompt when deleting reservations from the
admin-reservation component. The dialog prompt is a generic component
that can be used for any confirmation prompts. This prompt serves to
prevent accidental deletion of reservations.

* Add ability to delete old reservations

This commit adds the ability to delete old reservations. This is so the
 old reservations tab can be cleaned. However, this is a currently
hardcoded limit of 30 days that will not be deleted. This is done so no
recent reservations are deleted in the event that records need to be
checked.

* Add toggle to desk availability

This commit adds a toggle to the admin desk view to allow admins to
toggle the availability of a desk. This will also remove any
reservations that are currently active for the desk. It also includes
a prompt to confirm the action. This toggle is if the desk for any
reason is no longer available for use but should still be kept in the
system such as if it is broken and needs to be repaired.

* Add confirmation when user cancels reservation

Adds a confirmation prompt when a user cancels a reservation on the
Reservation page. This is to prevent accidental cancellations and so
users can be sure they are cancelling the correct desk and at the
correct time.

* Remove unnecessary imports

* Change how reservation infromation is refreshed

Added extra check for reserving desks on the same day to eliminate the
times that have already passed.
Added confirrmation prompt when removing a desk.
Reservation infromation no long causes a refresh and instead calls just
calls to get the data from the server for the specific part that has
changed. This is to prevent the page from jumping around when something
is changed.
Toggle availability was changed from a post to put request to be more
accurate.

* Delete reloadPage and delegate to functions

Removed reloadPage from the admin-reservation component and instead call
and load the data for the component that needs to be reloaded. This is
so that the position on the page will be saved and will not cause a full
refresh of the page.

* Remove unused code and adjust filter

list_all_desk_reservations_for_admin() is not used anymore, so it can be
removed.
make_desk_available() and make_desk_unavailable() are not used anymore
as their functionality is now covered by toggle_desk_availability().
Related tests are removed as well.
The filter was changed to block reservations for the current day if the
only times available are in the past.

* Change permissions on desk and desk_reservation

+++ frontend/src/app/reservation/reservation.component.ts
@@ -8,6 +8,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatDialog } from '@angular/material/dialog';
 import { ReservationDialogComponent } from '../reservation-dialog/reservation-dialog.component';
 import { Router } from '@angular/router';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 
 export interface DialogData {
@@ -100,24 +101,28 @@ export class ReservationComponent implements OnInit {
    * 
    */
   removeDeskReservation(deskReservationsListItem: [DeskReservation, Desk]): void {
+    let resDate = new Date(deskReservationsListItem[0].date)
+    let date = resDate.toLocaleDateString('en-US', {
+      weekday: 'long',
+      month: 'long',
+      day: 'numeric',
+      year: 'numeric',
+      hour: '2-digit',
+      minute: '2-digit'
+    })
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: { message:  "Are you sure you want to remove the reservation for " + deskReservationsListItem[1].tag + " on " + date + "?" },
+    })
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.deskReservationService.removeDeskReservation(deskReservationsListItem[1], deskReservationsListItem[0]).subscribe();
 
-    let reservationDate = new Date(deskReservationsListItem[0].date);
-    let reservationTime = reservationDate.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
+        let message = `Desk Reservation on ${date} canceled`;
+        this.snackBar.open(message, "", { duration: 4000 });
 
-    let formattedDate = reservationDate.toLocaleDateString('en-US',
-      {
-        weekday: 'long',
-        month: 'long',
-        day: 'numeric',
-        year: 'numeric'
-      })
-
-    this.deskReservationService.removeDeskReservation(deskReservationsListItem[1], deskReservationsListItem[0]).subscribe();
-
-    let message = `Desk Reservation on ${formattedDate} at ${reservationTime} canceled`;
-    this.snackBar.open(message, "", { duration: 4000 });
-
-    this.reloadPage();
+        this.reloadPage();
+      }
+    })
   }
